move controller to in.rest package
