+++ src/main/java/com/koc/user/adapter/in/rest/UserController.java
@@ -1,4 +1,4 @@
-package com.koc.user.application.controller;
+package com.koc.user.adapter.in.rest;
 
 import com.koc.user.application.dto.UserDto;
 import com.koc.user.application.service.UserService;
